Merge branch ADOdotNET-Integration Into Develop

+++ Producto.cs
@@ -2,12 +2,12 @@
 {
     public class Producto
     {
-        private uint _id;
+        private long _id;
         private string _descripcion;
-        private float _costo;
-        private float _precioVenta;
+        private decimal _costo;
+        private decimal _precioVenta;
         private int _stock;
-        private uint _idUsuario;
+        private long _idUsuario;
 
         public Producto()
         {
@@ -19,7 +19,7 @@
             this._idUsuario = 0;
         }
 
-        public Producto(uint id, string descripcion, float costo, float precioVenta, int stock, uint idUsuario)
+        public Producto(long id, string descripcion, decimal costo, decimal precioVenta, int stock, long idUsuario)
         {
             this._id = id;
             this._descripcion = descripcion;
@@ -29,11 +29,11 @@
             this._idUsuario = idUsuario;
         }
 
-        public uint Id { get => _id; set => _id = value; }
+        public long Id { get => _id; set => _id = value; }
         public string Descripcion { get => _descripcion; set => _descripcion = value; }
-        public float Costo { get => _costo; set => _costo = value; }
-        public float PrecioVenta { get => _precioVenta; set => _precioVenta = value; }
+        public decimal Costo { get => _costo; set => _costo = value; }
+        public decimal PrecioVenta { get => _precioVenta; set => _precioVenta = value; }
         public int Stock { get => _stock; set => _stock = value; }
-        public uint IdUsuario { get => _idUsuario; set => _idUsuario = value; }
+        public long IdUsuario { get => _idUsuario; set => _idUsuario = value; }
     }
 }
